+++ packages/python/m/log/config.py
@@ -1,6 +1,7 @@
 import logging
 
-from ..core.io import env
+from m.core import io as mio
+
 from .formatters import CiFormatter, JsonFormatter
 from .handlers import JsonFileHandler, StdErrHandler, StdOutHandler
 
@@ -20,7 +21,7 @@ def logging_config(level: int | None = None, json_file: str = '') -> None:
         json_formatter = JsonFormatter()
         all_handlers.append(JsonFileHandler(json_formatter, json_file))
 
-    debug_logs = env('DEBUG_M_LOGS', 'false') == 'true'
+    debug_logs = mio.env('DEBUG_M_LOGS', 'false') == 'true'
     default_level = logging.DEBUG if debug_logs else logging.INFO
     logging_level = level if level is not None else default_level
     logging.basicConfig(